move over userRepo
